+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-Overlay: blurred
-
 3.2.0
+
+- Pagination: Fixed truncation for total < 8
+- PinField, source: Updated keyboard tests
+- Flyout: Correctly calculate position for scaled down triggers
+- Flyout: Using vanilla event listeners on trigger, you can pass additional custom handlers yourself